refactor(skeleton): split item files for clarity
