+++ packages/components/skeleton/src/SkeletonRectangle.styles.ts
@@ -1,0 +1,13 @@
+import { cva } from 'class-variance-authority'
+
+export const skeletonRectangleStyles = cva(['flex'], {
+  variants: {
+    isLoading: {
+      true: ['bg-neutral/dim-4', 'min-h-lg min-w-lg', 'rounded-sm'],
+      false: [],
+    },
+  },
+  defaultVariants: {
+    isLoading: true,
+  },
+})
